Minor fixes for MIDCET-5076
In R47P0, MIDCET-5076 had a few issues. These are fixes on top of the
upstream driver.

Bug: 315967882
Test: Compile & Boot

+++ mali_kbase/mmu/mali_kbase_mmu_hw_direct.c
@@ -215,7 +215,7 @@ static int write_cmd(struct kbase_device *kbdev, unsigned int as_nr, u32 cmd)
 #if MALI_USE_CSF
 static int wait_l2_power_trans_complete(struct kbase_device *kbdev)
 {
-	u32 val;
+	u64 val;
 	const u32 timeout_us =
 		kbase_get_timeout_ms(kbdev, MMU_AS_INACTIVE_WAIT_TIMEOUT) * USEC_PER_MSEC;
 	const int err = kbase_reg_poll64_timeout(kbdev, GPU_CONTROL_ENUM(L2_PWRTRANS), val,
